Move TransactionFetcher to own file
